+++ src/main/java/com/piseth/java/school/phoneshop/repository/ColorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.piseth.java.school.phoneshop.model.Color;
 
-public interface ColorRepository extends JpaRepository<Color, Integer> {
+public interface ColorRepository extends JpaRepository<Color, Long> {
 
 }
